broker: prep for module class access restriction
Problem: the built-in module services call module_get_name() and
module_get_uuid() from the module thread context.

As a first step in an effort to restrict access to the module class
to only the main broker thread, change the module services to use the
flux::name and flux::uuid aux entries defined by RFC 5 instead.

Move the initialization of the aux entries from modservice_register()
the module thread itself.