flux-fsck: support index inputs to --rootref
Problem: Flux-fsck only allows you to fsck the current checkpoint.
However, there is a potential history of checkpoints stored that users
may wish to fsck.

Support the ability for users to specify checkpoint indices
to the --rootref option in flux-fsck.  This allows users to fsck
other checkpoints stored in the checkpoint history.