t: remove double checkpoint in fsck tests
Problem: In the flux-fsck tests in t2816-fsck-cmd.t, there are
several instances in which flux kvs sync is called right before
the kvs module is unloaded.  The KVS sync call is unnecessary since
the KVS module checkpoints when it is removed.

Remove the excess calls to "flux kvs sync".