terminus/pty/test: skip write tests on apple
problem: Macos ptys do not act like linux ptys. I've tried a variety of
things, but as far as I can tell any attempt to write to a leader fd
while no client has opened the follower file of the pty will result in
infinite EAGAIN errors from the write.  The pty test does not actually
open the follower side before writing to it.  I tried opening it in the
`flux_pty_open` call (as openpty normally does) which allows this test
to pass in both linux and macos, but then causes hangs in the attach
test in /t/ whether I close the fd there, or actually use the follower
fd in the attach rather than opening the follower file, several of these
could fix either the pty test, or the attach test, none of them could
fix both on both platforms.

solution: Skip the four sub-tests in test_pty which attempt to write to
an unusable leader pty file descriptor on macos until we can figure out
how to fix this more generally.