broker: convert groups to a built-in module
Problem: the broker uses the groups service for quorum
management, but groups is a standalone service that doesn't
need to be as tightly coupled to the broker as it is.

Convert groups to a built-in broker module.  The reduced coupling
between broker and groups should make both easier to maintain.