flux-fsck: support --scan option
Problem: Flux-fsck only allows you to fsck one root reference
or checkpoint at a time.  It would be useful for fsck to find
the most recent checkpoint that is valid.

Support a new --scan option that will iterate through all available
checkpoints and find the most recent one that passes.