broker: add autoload flag for builtin modules
Problem: all built-in broker modules are unconditionally loaded,
but it could be helpful to compile in modules that are conditionally
loaded to avoid the DSO path search overhead.

Add an 'autoload' flag to struct module_builtin and don't automatically
start modules that have autoload=false.  Those module can be loaded
with 'flux module load' as usual, but without the DSO overhead.