add libmodule_builtin.la convenience lib
Problem: making each built-in broker module a standalone
convenience library will require the broker Makefile.am
to be edited each time one is added.

Instead of building each module as a convenience library, build
one convenience library for all the built-in modules called
libmodule_builtin.la and link the broker with that.