ci: run tmate when dispatching workflow with debug set
Problem: Sometimes failures in ci are only reproducible on the github
action runners. To debug, developers have to modify the workflow
config to add the tmate action, then push changes to trigger a ci
build with tmate enabled.

To make this process easier, permanently add the tmate action to the
flux-core workflow config, but only have it trigger on a workflow
dispatch with debug enabled. This allows a developer to simply
check the "debug" box on the github actions page when restarting the
failing checks.

See https://github.com/mxschmitt/action-tmate?tab=readme-ov-file#manually-triggered-debug