doc: update tmate documentation
Problem: The tmate action in flux-core can now be triggered by
re-running jobs with the debug flag set, but documentation doesn't
mention this.

Update the docs to describe how to enable the tmate action. Keep the
diff that shows how the tmate action can be temporarily added to a
workflow for potential use in other framework projects.