job-manager: rename lost+found directory
Problem: In the near future flux-fsck will repair data and move
it to a lost+found directory.  This conflicts with the lost+found
directory used in the job-manager.

Rename the lost+found directory in the job-manager to job-lost+found.

Update one test in t/t2219-job-manager-restart.t for directory name
change.