ci: fix docker manifest generation
Problem: The set of images for which docker manifests are generated
is hand edited in the workflows file and so can easily become outdated.

Add a script to run the docker manifest commands based on current
output from `src/test/generate-matrix.py`, which contains the exact
set of docker tags for each image which then must be combined into
a manifest.