ci: add missing checkout step to generate-manifest
Problem: Docker manifest generation was moved to use scripts from
the flux-core repo, but a checkout of the repo isn't done in this job.

Add a checkout step so the necessary scripts are available. Also,
now that there are multiple steps in the job, make the whole
generate-manifest job conditional on tags or master.