modprobe: don't ignore failures from context.bash()
Problem: The Context class bash() helper function doesn't catch nonzero
exit codes from the subprocess, which could lead to an instance being
started in an inconsistent state.

Switch to subprocess.run() instead of Popen() and check the
CompletedProcess returncode. Raise an appropriate RuntimError if the
process exited with a non-zero status.