shell: add sysmon plugin
Problem: currently there is no built-in monitoring of cpu
and memory utilization for jobs.

Centrally collecting monitoring data for large jobs is a non-trivial
scalability challenge, but in the absence of that, we can easily
monitor utilization locally on each node (shell) and log significant
changes over time.

Add a shell plugin that uses the new sysmon service to monitor
for significant changes in the utilization of cpus allocated to job
and global memory utilization.  After an initial settling period, if
a usage change exceeds a threshold, the new value is logged in the
shell output.

The following options are supported:

  -o sysmon
      The plugin is disabled by default.  This (or any other
      valid sysmon option) enables it.

  -o sysmon.full
      Break down CPU utilization by CPU.  The default is to monitor
      combined CPU utilization as one value.

  -o sysmon.settle=FSD
      Suppress logging for the specified period after local tasks
      are started to avoid alerts during startup.  Default 10s.

  -o sysmon.mem-alert=PCT
      Suppress memory alerts for changes less than the specified
      percentage.  Setting it to 100 disables memory alerts.
      Default 20.

  -o sysmon.cpu-alert=PCT
      Suppress cpu alerts for changes less than the specified
      percentage.  Setting it to 100 disables cpu alerts.
      Default 20.

  -o verbose=1
      Each usage sample is logged as it is received.