shell: add sysmon plugin
Problem: currently there is no built-in monitoring of cpu
and memory utilization for jobs.

Add a shell plugin that samples data from the cgroups V2 cpu and
memory controllers.  The plugin is disabled by default, and can be
enabled with the 'sysmon' shell option.

The default behavior is to log the summary data from each shell
at the end of the job:

5.065s: flux-shell[1]: sysmon: memory.peak=3.1179504G
5.066s: flux-shell[1]: sysmon: loadavg-overall=2.00
5.068s: flux-shell[0]: sysmon: memory.peak=3.1179504G
5.068s: flux-shell[0]: sysmon: loadavg-overall=2.00

The plugin polls cgroups periodically.  By default, this is tied
to the Flux heartbeat event.  A different period may be selected by
setting the sysmon.period=FSD shell option.  Each polling cycle logs
sample information at trace level (view with -o verbose=2):

2.497s: flux-shell[0]: TRACE: sysmon: memory.current=1.3986053G
2.497s: flux-shell[0]: TRACE: sysmon: loadavg=2.00
2.497s: flux-shell[1]: TRACE: sysmon: memory.current=1.3990936G
2.497s: flux-shell[1]: TRACE: sysmon: loadavg=2.00

In this case the reported load average only covers the polling period.