modprobe: skip loading existing modules in `Modprobe.load()`
Problem: `Modprobe.load()` tries to load target modules and all their
dependencies, which results in unnecessary errors when modules are
already loaded.

Get the list of loaded modules and filter them from the set of target
modules to load. If the set of modules to load is empty after this
step, raise `FileExistsError` with an appropriate message that there
was nothing to do.