flux-fsck: keep track of missing valref references
Problem: In the near future we will want to repair corrupted
valref treeobjs.  In order to do so efficiently, we would like to
"save" the location of invalid indexes while we are fsck-ing
the valref treeobj.

Save the indexes of the missing references in a special
missing_indexes list.  This list is presently unused.