broker: add broker_panic()
Problem: there isn't a simple way to tell the broker that
a fatal error has occurred once it is up and running.

Add broker_panic() which logs a message at LOG_CRIT level,
then posts the "panic" event to the broker state machine
and sets the exit_rc to indicate that an error occurred.
When the panic event is posted to the state machine in
INIT, QUORUM, or RUN states, it transitions to the SHUTDOWN
state.