broker: panic if a module fails
Problem: if a module exits with a nonzero exit code AND module
unload was not requested, this condition is only logged so it
might slip by in tests.

Call broker_panic() when this occurs, unless broker.module-nopanic
is set to a value other than 0.

The panic causes the broker to shut down and exit with a nonzero
exit code.

Fixes #6962