doc: add note about `--signal=SIG@TIME` behavior in batch jobs
Problem: A signal sent to a batch job is delivered both to all the
running jobs and the batch script. The latter may be surprising or
catch users off guard if they expect only the jobs to be signaled.

Add a note to the submission commands documentation for `--signal`
describing signal behavior for batch jobs, and suggesting batch
scripts use a `trap` to capture `SIG` to avoid early exit.

Fixes #6964