broker: synchronize loading of builtin modules
Problem: there is no synchronization that ensures connector-local
is loaded before rc1 executes, so although it usually is, large
startup delays such as may be encountered in CI may cause problems.

There is a backoff-retry in the client to work around this but it
does not wait very long.  For the CI case, we extend the number of
retries by an environment variable in rc1.

Change modhash_load_builtins() to return a future that is fulfilled
once the autoload=true builtin modules have reached RUNNING state,
or is fulfilled with an error if any of them fail.

Change the broker state machine so that:
1. it is kicked off by calling a new function state_machine_kickoff()
instead of posting a "start" event that transitions from NONE->JOIN
2. state_machine_kickoff() calls modhash_load_builtins()
3. a continuation posts "builtins-success" to transition from NONE->JOIN

Although this ensures that builtins are loaded before INIT (rc1) state,
it doesn't delay any JOIN actions--there aren't any.  JOIN forces the
local broker to pause until the upstream broker is connected and has
reached RUN state, but the connect and the monitoring request are started
before the state machine is kicked off.

Update state machine sharness test due to the expected starting event
changing from "init" to "builtins-success".