broker: respond to unroutable internal requests
Problem: if the broker sends an RPC to a local service
that is not loaded, an ENOSYS response is not generated.

This is vestigial code from early Flux times when the
broker's internal flux_t handle was not entirely reactive.
It is now, so generate responses on error and avoid surprises
that may confound future explorers.