testsuite: require bash in script with bashisms
Problem: one of the issue tests uses 'let' in a generated
script, a bashism, but the script has #!/bin/sh.

This doesn't work on systems that link /bin/sh to dash(1),
although that doesn't seem to cause the test to fail.
(I just happened to notice the shell complaining on stderr).

Generate the script with #!/bin/bash.