etc: add script for running prolog,epilog,housekeeping
Problem: There is no location that allows other packages on a Flux
system to install scripts that will be run as part of the job prolog,
epilog or housekeeping when these are configured in a system instance.
This is because the current implementation runs a single site-provided
script in `/etc/flux/system/{prolog,epilog,housekeeping}`, which has
no guarantee of running scripts from a standard `.d` directory.

Add a new script installed to `$libexecdir/flux/flux-run-system-scripts`
which is meant to be run as a wrapper for the system prolog, epilog,
and housekeeping scripts. This new script runs system and site-provided
scripts in the following order:

 1. package provided scripts in `$libexecdir/flux/<name>.d/*`

 2. For backwards compatibility, if `/etc/flux/system/<name>` exists,
    then this script is executed, otherwise, run all site-provided
    scripts in `/etc/flux/system/<name>.d/*`

Where `<name>` is provided as the first and only argument to
`flux-run-system-scripts`, e.g. `flux-run-system-scripts prolog`.

By default, `flux-run-system-scripts` will exit with the largest
script exit code after running all discovered scripts. This behavior
can be changed on a case-by-case basis via the broker config:

 `job-manager.<name>.exit-on-first-error`

If this config value is `true`, then `flux-run-system-scripts` will
exit on the first error from any script.