testsuite: run cleanup between tests in t9000-system.t
Problem: System tests under `t/system` that use `cleanup()` assume
that the cleanup will occur before the next test file runs, but this
is not the case because of the way t9000-system.t works. Instead,
cleanup is deferred until the end of all tests.

Force cleanup to run between test files.