testsuite: add system tests of job prolog
Problem: There are no system tests that exercise the job prolog running
as a systemd unit, as would be configured for a Flux system instance.
Particularly, the new flux-run-system-scripts script is not tested.

Add t/system/0006-perilog.t with a simple sanity check of this feature
using the job prolog.