job-manager: support `*.exit-on-first-error` config
Problem: The `job-manager.{name}.exit-on-first-error` where name is
prolog, epilog, or housekeeping is used by the flux-run-system-scripts
script to determine how to handle individual script errors. However,
attempting to set these config keys in the broker results in an error
because the job-manager and perilog plugin reject unknown keys.

Add `exit-on-first-error` to the list of keys unpacked in housekeeping
and the perilog plugin so that these config keys can be set.