resource: watch for R updates even with resource.rediscover
Problem: When resource.rediscover=true, the KVS watch of the job's R
is canceled after the first response. This causes updates of a job's
expiration to be ignored by instances started with this config key.

Keep the R watch future active even with resource.rediscover so that
updates to the job's expiration are fully supported in this mode.

Fixes #6984