python: fix tls initialization in flux.conf_builtin
Problem: The `tls.FLUX_CONF_AUTO_FLAG` attribute is apparently not
initialized properly. Possibly it is only initialized in the main
thread. This results in an `AttributeError` raised when using
`flux.conf_builtin_get()` from a thread other than the main thread.

Instead of initializing this attribute in the main module to `None`,
just check for the attribute when initializing using `hasattr()`.