doc: update description of rc1_path & rc3_path
Problem: The broker.rc1_path and broker.rc3_path attributes and
rc1_path, rc3_path config keys are described as a path to the rc1 and
rc3 scripts used by the broker, but now these attributes and config
are used more generically as a command for the broker to execute in
these runlevels.

Update the documentation of these attrs and builtin config.