testsuite: fix spurious heartbeat errors in t2801-top-cmd.t
Problem: The heartbeat module is removed from rank 0 in
t2801-top-cmd.t, but left active on the other ranks in the test
instance. This causes child brokers to disconnect and exit if the
test runs longer than the default heartbeat timeout window, 5 minutes.

Remove the heartbeat module on the other broker ranks as well.