flux-fsck: output directories in verbose mode
Problem: Directories being fscked are not output under verbose
mode.  This is an issue especially if the directory is empty, thus
the fsck of the directory is never known to the user.

Output the path being fscked on all keys, including directories.

Update tests in t2816-fsck-cmd.t to include the extra expected
output in tests.