testsuite: fix flaky shell-lost test
Problem: t/issues/t2492-shell-lost.sh is flaky in ci. This is likely
due to not waiting long enough for the shell lost message in the test,
which could result in sending SIGINT to job shells before the job
script has actually started.

Extend the timeout in the `flux job wait-event` call. Also, improve
the error on failure by terminating the test immediately if the
wait-event call fails.