broker: create default statedir in /var/tmp
Problem: if statedir is not set on the command line, KVS content
is stored in rundir which defaults to /tmp, but /tmp is inappropriate
for large files.

Instead of making statedir optional, treat it the same as rundir.
That is, if it is not specified on the command line, create it
and schedule it for cleanup.  Use /var/tmp if TMPDIR is not set.
The statedir is only created on rank 0, and the statedir attribute
is now unset on other ranks.

Fixes #6996