broker: add {statedir,rundir}-cleanup attrs
Problem: the presence of the statedir attribute is used as a
flag by content-sqlite to select database flags appropriate
for persistence, but statedir is now always set.

Add a statedir-cleanup attribute to signify whether statedir
is to persist or not.  Add rundir-cleanup also for symmetry.
Allow these to be set on the command line to override the
default cleanup behavior.