t: fix race in t2315
Problem: there appears to be a race between the system resource
configuration and the tests querying the available resources

Use flux run to wrap the query such that the resources must be "up"
in order to be allocated by the scheduler. Other queries on state "all"
should be fine as resources will still be seen even if not yet "up".

For "bad-exclude" and "bad-host-exclude" the query is changed to match
"conftest-bad" and all of them to query "all" resources to make clear
that the output is only for debugging (previous nnodes file was empty).