modprobe: activate required modules of selected alternatives
Problem: If a module is an alternative provider of a service, and the
module is selected as the highest priority alternative, any required
modules of that module are not also activated. If the alternative
module also `needs` the required module, then the selected alternative
will be subsequently disabled since the module it needs is not enabled.

When adding a module to the current set of active tasks, also add
all its required modules.