libeventlog: cap eventlogger kvs txn ops
Problem: The number of operations in a kvs txn generated by the
eventlogger has no cap.  In the future when the KVS may have a
transaction operations cap, this could cause problems.

If the number of operations in a eventlogger batch exceeds a cap,
initiate the transaction immediately, even if a timer has not
been reached.

For the time being, set the cap to a hard coded value.  This may
become configurable in the future.

Fixes #6910