testsuite: fix barrier in shell-lost issues test
Problem: The `flux pmi barrier` call in t2492-shell-lost.sh is not
functional because Python's `subprocess.run()` is closing the PMI_FD.
This causes the barrier to exit prematurely and results in a race where
the test job can get killed by SIGINT before it is handling the signal.

Add `close_fds=False` to the `subprocess.run()` call so that the PMI
fd is not closed and the barrier functions properly.