broker: interpret sticky bit in -Sstatedir=DIR
Problem: there is not a simple way to use an alternate
/tmp like directory for ONLY the statedir in the current
instance, since setting TMPDIR both statedir and rundir
and is inherited by children of flux.

If statedir/rundir is set on the command line and the
specified directory has the sticky bit set, create a temporary
directory within that directory rather than trying to use it
as the rundir/statedir directly.

Fixes #6990