modprobe: fix get_deps() with module provides list
Problem: When calculating the precedence list of tasks in
`Modprobe.get_deps()`, modules listed by service name are dropped
if a module providing that service name was provided in the `tasks`
argument instead of the service name.

Extend the tasks set by the `provides` list from all target arguments
so that the test for existence of either the explicit module or its
provided service name(s) succeeds later in the function.