modprobe: resolve alternatives corner cases in modprobe
Problem: There are a couple corner cases in modprobe revolving
around the use of alternatives.

 1. When explicitly loading a module that is not the current
    highest-priority alternative, the dependency resolution may pull
    in the higher priority alternatives, resulting in errors when
    conflicting modules are loaded.

 2. When unloading modules, a similar situation occurs if the
    current highest-priority module is not loaded.

Add a new function that ensures currently alternatives are all set
to a given list of modules. Use this to force the alternatives to
currently loaded modules in the case of `remove`, or the explicitly
requested modules in `load`.