modprobe: fix modprobe module config update logic
Problem: When a module is updated via config, the update is done
using the raw TOML values, but some module attributes, like ranks,
should be objects.

Create an instance of a Module task (or allow one to be passed
in to the function) before updating the existing task so that all
updated attributes are of the proper type. Rename `update_task()`
to `update_module()` to make it clear this function now only works
for modules.