testsuite: try harder to cleanup old units in t2412-sdmon.t
Problem: If a systemd unit created by a past invocation of
t2412-sdmon.t is in a failed state, then `systemctl -user stop` does
not always clean up the old unit, which causes some of the following
tests to fail.

Try `systemctl --user reset-failed` in addition to `stop` on the
target units of the test to ensure better cleanup.