libeventlog: set reactor when creating future
Problem: In the rare case when a eventlogger batch is committed
but there is nothing to commit, an "empty" fulfilled future
is returned.  However, the reactor is not set in this future.  This
can lead to errors if callers then try to pass this future to
functions like flux_future_then().

Set the reactor after creating the fulfilled future.

Fixes #7028