libutil: add cgroup_path_to() drop cgroup_access()
Problem: there is no way to use the cgroup helpers to generate
a path to a cgroup file for use with inotify(7).

Add cgroup_path_to() which uses a buffer in thread-local storage
to build a path to any cgroup file.

Drop cgroup_access(), which is just a wrapper around access(2)
that builds the path.  Such wrappers no longer save much code.

Update the cgroup unit test and the sysmon shell plugin.