flux-cgroup: add command to print cgroup path
Problem: finding the cgroup path for the current process
is a bit tricky on different systems, making it hard to write
portable tests or review cgroup settings on a live system.

Add 'flux cgroup path' which uses well tested C code to find
the cgroup path for the current process.