modprobe: add debug logging during conf/rc file loading
Problem: There are no debug logs emitted from modprobe when loading
modprobe TOML and rc Python files, but this may be useful now that
the search strategy is more complicated.

Add a `Modprobe.print()` function that just wraps `Context.print()`
for use in verbose logging.

Log messages about what paths are being searched and the actual
files that are loaded for both TOML and Python files when a Modprobe
instance is in verbose mode.