broker: send name and uuid in module welcome msg
Problem: module threads receive name and uuid as mod_main()
parameters, but that is inconvenient for modules run as
separate processes.

Drop the module name and uuid from mod_main() arguments
and add them to the module welcome message.

Add the interthread uri to the mod_main() arguments.
The uri, which is required to connect to the broker and receive
the welcome message, was formerly constructed from the passed uuid.

Adjust the flow of control in mod_main() so the welcome message is
received before the name or uuid are used.  This drops the module
name from some early log messages, but since those are both unlikely
and not coming from module-specific code, perhaps that is OK.