flux-jobs: show full usernames by default
Problem: the default format of flux-jobs(1) cuts usernames off at
8 characters. This is a bit of an outdated default. Usernames can
be much longer, and now frequently are, and users work around this
all the time.

Update the default to be a floor of 8 characters, with the cutoff
set at the longest username's length. Update the other core-
provided formats in the same way to be consistent.