broker: suppress module.status ENOENT error
Problem: when a broker module is loaded in test outside of
the broker, an error occurs when internal services are
registered or when flux_module_set_running() is called.

Since the module is simply not in the module hash in this
case, the request to update the broker's record of its status
does not apply.  Treat this as a no-op rather than an error.

Update a couple tests in t0003-module.t to expect this.