testsuite: fix a couple tests in t0100-modprobe.t
Problem: A couple of tests in t0100-modprobe.t use `libexecdir`
as the expected path to `modprobe.d`, but this should be `datadir`.
This works for the build tree since `libexecdir` == `datadir`, but
will fail when testing installed Flux. Additionally, use of `$(pwd)`
as a search path entry may not be failsafe, since there is a chance
a different path could be printed by modprobe, e.g. in the case of
a symlink.

Fix these issues by using `datadir` for expected paths to modprobe
TOML files, and an explicit fake path in place of `$(pwd)` when
matching exact paths in output.

Additionally, add more useful debug output to these tests.