flux-jobs: use "+" for multi-line help message
Problem: The help message for the -u/--user optional argument does not
use a "+" sign to attach the second line like the other multi-line help
messages in flux-jobs.py.

Add a "+" sign before the second line of the help message to make it
consistent with the other multi-line help messages.