python: consider output header w/ expandable width
Problem: The expandable field prefix ("+:") does not consider
the width of the output header name when calculating output width.
If all output entries are less than the size of the output header
name, then output will not look as expected.

Solution: When calculating maximum output width of a field, take
into account the header entry name as long as no_header is not set.

Update some tests in python/t0024-util.py for changes.

Fixes #7030