doc: add note about expanded width with precision
Problem: In flux-jobs(1), if the expanded field width modifier
("+:") is used on a field with a precision, the precision is also
updated to the maximum width.  This is not documented.

Add note that both the width and precision of a field will be updated
by the expanded field width modifier.