python: support no_header in OutputFormat filter()
Problem: In the future the header may be needed in calculations
within the OutputFormat filter() function.  There is no way to
inform the filter() function that the header should not be considered
(for example, if the header fill not be output).

Add a no_header flag to the filter() function in OutputFormat().  This
parameter currently does nothing.

Update all callers as needed.