python: support line truncation in `OutputFormat` class
Problem: Users of the `OutputFormat` class do not have a way to
truncate lines of output, e.g. to the current terminal width.

Add a `width` parameter to both the `print_items()` and `header()`
methods of `OutputFormat` which, when set > 0, will truncate all
output lines to the given number of columns. If set to 0, width will
default to the current terminal size if a terminal is found.