libflux: fix flux_pollfd() file descriptor leak
Problem: calling flux_pollfd() causes an epoll file descriptor
to be created internally, but the file descriptor is not closed
when the flux_t handle is destroyed.

Since each call to flux_pollfd() reuses the internal file descriptor,
it is clearly "owned" by the flux_t handle.  Close it when the
handle is destroyed.