flux_pollevents(3): update man page
Problem: flux_pollevents(3) does not mention that the pollfd
file descriptor is the property of the flux_t handle and includes
a big cut and paste example without explaining the simple concepts
it was meant to elucidate.

Mention the rules of the returned pollfd.
Replace the example with a bit of explanation.