libflux: do not require out parameters
Problem: In flux_response_decode_raw() the data and length out parameters
are required.  This is not consistent with most functions of this style,
as out parameters are typically optional.  For example, perhaps a caller
is only interested in the length of the data returned and not the data
itself.

Do not require the data or len out parameters to flux_response_decode_raw()
to be required.

Fixes #6824