libflux: fix flux_pollevents() POLLOUT logic
Problem: flux_pollevents() always raises FLUX_POLLOUT,
even when the underlying connector is not ready for writing.

The connector pollevents bits are ORed with those from the
re-queue deque, but the re-queue deque should only be contributing
POLLIN and POLLERR, not POLLOUT, as it cannot sink messages from
flux_send().

Ignore the re-queue deque's POLLOUT bit in flux_pollevents().