python: update docs on task 'requires' argument
Problem: The 'requires' argument for a modprobe task only suggests
that the 'requires' arguments may not be run "before" any of its
requirements.  But it is really "before" or "after".

Add note that 'requires' arguments do not imply running before or
after the task.  This informs the user that 'requires' arguments
could run in parallel.