libflux: make msg_deque_pollfd more readable
Problem: for some reason, msg_deque_pollfd() initializes
the eventfd(2) counter to the pollevents mask rather than
1 or 0.

This should have no effect since EFD_SEMAPHORE is not in effect,
which means that a read always sets the counter to zero,
but it makes the intent somewhat confusing.

Use the event flag value instead.