libflux: reduce spurious pollfd wake-ups (case 2)
Problem: each time a POLLIN event is handled on a loop://
or interthread:// handle, a spurious POLLIN is raised.

This is because of the msg_deque_pollfd() behavior just documented,
where POLLOUT is raised when msg_deque_pollevents() is called.
The flux_t handle monitors POLLIN | POLLOUT on the connector fd
because it doesn't know how the connector is implemented.
Hence the spurious wake-up for those connectors.

Add a connector option that can be interrogated by flux_pollfd()
to determine which events should be monitored on the connector
implementation's pollfd.  If this the option is not implemented,
monitor POLLIN | POLLOUT as before.

Implement this option in the loop:// and interthread:// connectors.

Fixes #7067