libflux: reduce spurious pollfd wakeups (case 1)
Problem: each time a POLLIN event is handled and flux_pollevents()
is called, a spurious POLLIN is raised.

This is because of the msg_deque_pollfd() behavior just documented,
where POLLOUT is raised when msg_deque_pollevents() is called.
The flux_t handle monitors POLLIN | POLLOUT on the requeue deque,
but only POLLIN should be of interest.

Drop POLLOUT from the interest list.

This partially fixes #7067