broker: optionally load modules as processes
Problem: although Flux broker extension modules are only allowed
to interact with each other using messages, this isolation is not
enforced and a badly behaved module can crash the broker.

Add an 'exec' flag to the module.load RPC.  When set, the broker
starts 'flux broker-exec MODULE' as a sub-process and communicates
with it over a sub-process message channel.  The module is thereby
placed in an isolated address space where it can do less harm when
misbehaving.

Enable RPC tracking so if the process dies unexpectedly, error
responses are automatically sent to outstanding RPCs.  Modules
executing as threads do not need this capability since they are
expected to clean up on failure.

Some refactoring was needed to avoid code duplication among the
three module loading methods (builtin, thread, exec).