modprobe: add `exec` option to Module config
Problem: There is no support in modprobe for loading modules as
processes instead of threads.

Add a new `exec` key to the modprobe module configuration. When false
(the default), the module is loaded as thread. When true, the `exec`
flag is set in the `module.load` RPC payload which loads the module
as a separate process.

Fixes #7077