libsubprocess: add `flux_rexec_bg(3)`
Problem: There is no way to launch a subprocess and allow it to keep
running while the frontend tool such as flux-exec(1) exits.

Add a new `flux_rexec_bg(3)` API call to libsubprocess that issues
an exec RPC with a `SUBPROCESS_REXEC_BACKGROUND` flag (as defined in
RFC 42).

On the server side, the background flag terminates streaming RPC
responses after the process is successfully started or fails, but keeps
a successful process running until completion (or until the broker
is shut down). Output from the process is copied to the broker logs.