testsuite: simplify test to avoid hang
Problem: a test in t3203-instance-recovery.t hangs in CI.

I'm not sure what's going on but the test is just a cut and paste
of the one above when it could be simplified to avoid the need
for the runpty wrapper, which might be creating the environment
needed for the hang.

Simplify the test.

Not a fix to #7091 but maybe a workaround for now.