broker: disable duplicate buffering on rc subprocs
Problem: now that the broker rc scripts are run as a "remote"
sub-processes (in another thread), line buffering is handled
remotely and local buffer management, with its large buffer and
multiple watchers, is not needed locally.

Set the FLUX_SUBPROCESS_FLAGS_LOCAL_UNBUF flag on the runat
remote sub-processes and change the stdio callback to use
`flux_subprocess_read()` instead of `flux_subprocess_getline()`,
as required by that flag.