broker: simplify internal event routing
Problem: at some point we may want to offload publishing to
a separate thread, but way the broker currently processes
events is a maze of callbacks.

Simplify by moving event publishing to the overlay class.
The flow between broker and overlay is now

1) when an event message is received from a broker module or the
broker's internal flux_t handle, it is forwarded directly to the overlay
for publishing

2) when the overlay receives a published event, it is forwarded directly
to broker modules and the broker's internal flux_t handle (filtered by
subscriber list).

3) events received by the overlay from the broker are sent upstream to
rank 0, where they are assigned a sequence number and published,
propagating downstream.  Propagation is handled completely within the
overlay code.

4) the event.publish RPC is an alternate event publishing mechanism that
allows the client to obtain the event sequence number.  This is now
handled by the overlay on rank 0 instead of the publisher class.