broker: rename event.publish to overlay.publish
Problem: event.subscribe and event.unsubscribe are handled by
the broker, but event.publish is handled by the overlay code,
which might move to a built-in broker module at some point.

Change the RPC target from event.publish to overlay.publish
so the publish method can move outside of the broker with
the overlay code.

Update tests.